+++ playground/src/pages/misc/directives/DirectiveSubComponent.vue
@@ -1,7 +1,0 @@
-<script setup lang="ts">
-import { vLightHelper } from '@tresjs/core'
-</script>
-
-<template>
-  <TresDirectionalLight v-light-helper :position="[-1, 1, 0]" :intensity="1" />
-</template>