+++ playground/vue/src/pages/advanced/suspense/AsyncComponent.vue
@@ -1,0 +1,11 @@
+<script setup lang="ts">
+import { useTresContext } from '@tresjs/core'
+import { useGLTF } from '@tresjs/cientos'
+
+const { scene } = await useGLTF('https://raw.githubusercontent.com/Tresjs/assets/main/models/gltf/aku-aku/AkuAku.gltf', { draco: true })
+useTresContext().invalidate()
+</script>
+
+<template>
+  <primitive :object="scene" :position-y="-2" />
+</template>