+++ playground/vue/src/pages/loaders/fbx-loader/index.vue
@@ -1,0 +1,11 @@
+<script setup lang="ts">
+import { TresCanvas } from '@tresjs/core'
+
+import TheExperience from './TheExperience.vue'
+</script>
+
+<template>
+  <TresCanvas clear-color="#C0ffee">
+    <TheExperience />
+  </TresCanvas>
+</template>