+++ playground/vue/src/router/index.ts
@@ -1,0 +1,15 @@
+import { createRouter, createWebHistory } from 'vue-router'
+import { allRoutes } from './routes'
+
+const routes = [
+  {
+    path: '/',
+    name: 'Home',
+    component: () => import('../pages/index.vue'),
+  },
+  ...allRoutes,
+]
+export const router = createRouter({
+  history: createWebHistory(),
+  routes,
+})